requirements.txt
PySide6>=6.5.0
Pillow>=9.0.0
# opencv-python (optional, uncomment if needed for advanced processing)
# opencv-python>=4.5.0
